requirements.txt
chromadb
psycopg2-binary
python-dotenv
schedule
